17231: Moved redirect code to other lifecycle method
 Arvados-DCO-1.1-Signed-off-by: Daniel Kutyła <daniel.kutyla@contractors.roche.com>

+++ src/views-components/api-token/api-token.tsx
@@ -4,6 +4,7 @@
 
 import { RouteProps } from "react-router";
 import * as React from "react";
+import { RootState } from "~/store/store";
 import { connect, DispatchProp } from "react-redux";
 import { saveApiToken } from "~/store/auth/auth-action";
 import { getUrlParameter } from "~/common/url";
@@ -12,38 +13,42 @@ import { navigateToRootProject, navigateToLinkAccount } from "~/store/navigation
 import { Config } from "~/common/config";
 import { getAccountLinkData } from "~/store/link-account-panel/link-account-panel-actions";
 import { replace } from "react-router-redux";
+import { User } from "~/models/user";
 
 interface ApiTokenProps {
     authService: AuthService;
     config: Config;
     loadMainApp: boolean;
+    user?: User;
 }
 
-export const ApiToken = connect()(
+export const ApiToken = connect((state: RootState) => ({
+    user: state.auth.user,
+}), null)(
     class extends React.Component<ApiTokenProps & RouteProps & DispatchProp<any>, {}> {
         componentDidMount() {
             const search = this.props.location ? this.props.location.search : "";
             const apiToken = getUrlParameter(search, 'api_token');
-            const loadMainApp = this.props.loadMainApp;
-            this.props.dispatch<any>(saveApiToken(apiToken)).finally(() => {
-                const redirectURL = this.props.authService.getTargetURL();
-
-                setTimeout(() => {
-                    if (loadMainApp) {
-                        if (redirectURL) {
-                            this.props.authService.removeTargetURL();
-                            this.props.dispatch(replace(redirectURL));
-                        }
-                        else if (this.props.dispatch(getAccountLinkData())) {
-                            this.props.dispatch(navigateToLinkAccount);
-                        }
-                        else {
-                            this.props.dispatch(navigateToRootProject);
-                        }
-                    }
-                }, 0);
-            });
+            this.props.dispatch<any>(saveApiToken(apiToken));
         }
+
+        componentDidUpdate() {
+            const redirectURL = this.props.authService.getTargetURL();
+
+            if (this.props.loadMainApp && this.props.user) {
+                if (redirectURL) {
+                    this.props.authService.removeTargetURL();
+                    this.props.dispatch(replace(redirectURL));
+                }
+                else if (this.props.dispatch(getAccountLinkData())) {
+                    this.props.dispatch(navigateToLinkAccount);
+                }
+                else {
+                    this.props.dispatch(navigateToRootProject);
+                }
+            }
+        }
+
         render() {
             return <div />;
         }
